creditCard/ccf/requirement.txt
asgiref==3.3.1
cachetools==4.2.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cmake==3.18.4.post1
colorama==0.4.4
Django==3.1.7
django-credit-cards==0.4.1
django-crispy-forms==1.11.2
django-qrcode==0.3
dlib==19.18.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.10
ipregistry==3.0.0
joblib==1.0.1
numpy==1.20.2
opencv-python==4.5.1.48
pandas==1.2.3
Pillow==8.2.0
pyotp==2.6.0
python-dateutil==2.8.1
pytz==2021.1
qrcode==6.1
requests==2.25.0
scikit-learn==0.24.1
scipy==1.6.2
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
urllib3==1.26.4
